+++ contracts/kernel/FsLib1.sol
@@ -23,7 +23,7 @@ library FsLib1 {
     }
   }
 
-  function chmod(FsLib.Disk storage self, bytes calldata path, uint mode, uint curdir) external onlyOwner(self) {
+  function chmod(FsLib.Disk storage self, bytes calldata path, uint16 mode, uint curdir) external onlyOwner(self) {
     (uint ino,,) = self.pathToInode(path, curdir, 2);
     require(ino > 0, 'ENOENT');
     FsLib.Inode storage inode = self.inode[ino];
@@ -35,10 +35,10 @@ library FsLib1 {
     (uint ino,, bytes memory key) = self.pathToInode(source, curdir, 2);
     require(ino > 0, 'ENOENT');
     FsLib.Inode storage inode = self.inode[ino];
-    require(inode.fileType != FileSystem.FileType.Directory, 'EISDIR');
+    require(inode.fileType != uint8(FileSystem.FileType.Directory), 'EISDIR');
     (uint ino2, uint dirIno, bytes memory key2) = self.pathToInode(target, curdir, 2);
     if (ino2 > 0) {
-      require(self.inode[ino2].fileType == FileSystem.FileType.Directory, 'EEXIST');
+      require(self.inode[ino2].fileType == uint8(FileSystem.FileType.Directory), 'EEXIST');
       dirIno = ino2;
     } else {
       key = key2;
@@ -53,11 +53,11 @@ library FsLib1 {
     require(ino == 0, 'EEXIST');
     ino = self.allocInode();
     FsLib.Inode storage inode = self.inode[ino];
-    inode.fileType = FileSystem.FileType.Symlink;
+    inode.fileType = uint8(FileSystem.FileType.Symlink);
     inode.mode = 511;
     inode.links = 1;
     inode.owner = inode.group = tx.origin;
-    inode.lastModified = now;
+    inode.lastModified = uint64(now);
     uint inoExtent = self.allocInodeExtent();
     inode.data[''] = inoExtent;
     inode.keys.push('');
@@ -71,19 +71,19 @@ library FsLib1 {
     (uint ino,,) = self.pathToInode(path, curdir, 1);
     require(ino > 0, 'ENOENT');
     FsLib.Inode storage inode = self.inode[ino];
-    require(inode.fileType == FileSystem.FileType.Symlink, 'EINVAL');
+    require(inode.fileType == uint8(FileSystem.FileType.Symlink), 'EINVAL');
     return self.inodeExtent[inode.data['']].extent;
   }
 
   function move(FsLib.Disk storage self, bytes calldata sourcePath, bytes calldata targetPath, uint curdir) external onlyOwner(self) {
     FsLib.ResolvedPath memory source = self.pathToInode2(sourcePath, curdir, 1);
     require(source.ino > 0, 'ENOENT');
-    bool sourceIsDir = self.inode[source.ino].fileType == FileSystem.FileType.Directory;
+    bool sourceIsDir = self.inode[source.ino].fileType == uint8(FileSystem.FileType.Directory);
     FsLib.ResolvedPath memory target = self.pathToInode2(targetPath, curdir, 2);
     if (source.ino == target.ino) return;
     if (target.ino > 0) {
       FsLib.Inode storage inode = self.inode[target.ino];
-      if (inode.fileType == FileSystem.FileType.Directory) {
+      if (inode.fileType == uint8(FileSystem.FileType.Directory)) {
         target.dirIno = target.ino;
         target.key = source.key;
         if (target.dirIno == source.dirIno) return;
@@ -91,11 +91,11 @@ library FsLib1 {
           uint ino = self.inodeValue[inode.data[target.key]].value;
           inode = self.inode[ino];
           if (sourceIsDir) {
-            require(inode.fileType == FileSystem.FileType.Directory, 'ENOTDIR');
+            require(inode.fileType == uint8(FileSystem.FileType.Directory), 'ENOTDIR');
             require(inode.keys.length == 2, 'ENOTEMPTY');
             if (--inode.refCnt == 0) self.freeInode(ino);
           } else {
-            require(inode.fileType != FileSystem.FileType.Directory, 'EISDIR');
+            require(inode.fileType != uint8(FileSystem.FileType.Directory), 'EISDIR');
             if (--inode.links == 0) self.freeInode(ino);
           }
         }
@@ -123,7 +123,7 @@ library FsLib1 {
     FsLib.ResolvedPath memory source = self.pathToInode2(sourcePath, curdir, 2);
     require(source.ino > 0, 'ENOENT');
     FsLib.Inode storage inode = self.inode[source.ino];
-    bool sourceIsDir = inode.fileType == FileSystem.FileType.Directory;
+    bool sourceIsDir = inode.fileType == uint8(FileSystem.FileType.Directory);
     FsLib.ResolvedPath memory target = self.pathToInode2(targetPath, curdir, 2);
     if (source.ino == target.ino) return;
     if (sourceIsDir) {
@@ -141,11 +141,11 @@ library FsLib1 {
 
   function _copy(FsLib.Disk storage self, FsLib.ResolvedPath memory source, FsLib.ResolvedPath memory target) private {
     FsLib.Inode storage inode = self.inode[source.ino];
-    bool sourceIsDir = inode.fileType == FileSystem.FileType.Directory;
+    bool sourceIsDir = inode.fileType == uint8(FileSystem.FileType.Directory);
     uint newIno;
     if (target.ino > 0) {
       FsLib.Inode storage inode2 = self.inode[target.ino];
-      if (inode2.fileType == FileSystem.FileType.Directory) {
+      if (inode2.fileType == uint8(FileSystem.FileType.Directory)) {
         target.dirIno = target.ino;
         target.key = source.key;
         if (target.dirIno == source.dirIno) return;
@@ -154,11 +154,11 @@ library FsLib1 {
           uint ino = self.inodeValue[inode2.data[target.key]].value;
           inode2 = self.inode[ino];
           if (sourceIsDir) {
-            require(inode2.fileType == FileSystem.FileType.Directory, 'ENOTDIR');
+            require(inode2.fileType == uint8(FileSystem.FileType.Directory), 'ENOTDIR');
             self.checkMode(inode2, 3);
             newIno = ino;
           } else {
-            require(inode2.fileType != FileSystem.FileType.Directory, 'EISDIR');
+            require(inode2.fileType != uint8(FileSystem.FileType.Directory), 'EISDIR');
             if (--inode2.links == 0 && inode2.refCnt == 0) {
               newIno = ino;
             }
@@ -183,7 +183,7 @@ library FsLib1 {
   function copyInode(FsLib.Disk storage self, uint ino, uint ino2, uint dirIno) private {
     FsLib.Inode storage inode = self.inode[ino];
     FsLib.Inode storage inode2 = self.inode[ino2];
-    bool sourceIsDir = inode.fileType == FileSystem.FileType.Directory;
+    bool sourceIsDir = inode.fileType == uint8(FileSystem.FileType.Directory);
     self.checkMode(inode, sourceIsDir ? 5 : 4);
     if (!sourceIsDir || inode2.keys.length == 0) {
       inode2.fileType = inode.fileType;
